Improve signature help active parameter with context (#167)
Use `validateFunctionArguments` to match the cursor position to the correct parameter. This means that the signature will be able to highlight out-of-order arguments correctly, including `*args` and `**kwargs`.

Unfortunately, LSP isn't powerful enough to actually do this perfectly for Python. Only one signature may be marked active and have an active parameter, but as the user's typing multiple overloads may still apply and their parameters be in entirely different orders.

For now, settle for selecting the first overload containing the active parameter, which is right a lot of the time.