Missing parameter on extracting inside a for loop statement (#786)
* Missing parameter on extracting inside a for loop statement

The logic aroud finding the parentNode was flawed in the sense that it was going up too high, always going up to the function or module level.  In this test case "w" wasn't in the function scope so we weren't adding it as a parameter to the new extracted function.

Now i'm starting at the startNode then then going up the parent tree until we find a node that covers both startNode and endNode.

Also fixing logic issues where i was checking if a variable was being written/declared before being read.  Previously i was comparing the file offset, but that doesn't work for things happening on the same line.  Expressions further right of the selection or offset will execute before things on the left.

Fixes #779

* adding check to see if our parentNode is already a suite or module before tyring to find the suite or module of our parent node.  This should prevent returning our grandparent suite by mistake.

* eslint